Include react navigation component
