+++ components/LaunchPad/index.js
@@ -1,3 +1,2 @@
 import LaunchPad from './LaunchPad';
-
 export default LaunchPad;
